setup header ui basics
