+++ src/pages/components/MainGameContainer/MainGameContainer.module.scss
@@ -1,0 +1,34 @@
+@import "../../../styles/variables";
+
+.main_game_container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: $spacing-lg;
+}
+
+.game_header_container {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  background-color: $accent-color;
+  padding: 1rem 1rem;
+  border-radius: $border-radius-lg;
+  width: 100%;
+}
+
+.game_header_title {
+  font-size: 2.5rem; /* Larger font size */
+  font-weight: 700; /* Bold font weight */
+  text-shadow: 0px 4px 10px rgba(232, 67, 147, 0.7), 0px 0px 15px rgba(232, 67, 147, 1);
+  letter-spacing: 2px; /* Adds spacing between letters */
+  text-transform: uppercase; /* Makes the text uppercase */
+  transition: transform 0.3s, text-shadow 0.3s;
+}
+
+.circle {
+  width: 1rem;
+  height: 1rem;
+  background-color: $panel-color;
+  border-radius: 50%;
+}
