+++ src/pages/components/MainGameContainer/QuoteDisplay/QuoteDisplay.module.scss
@@ -4,7 +4,7 @@
   display: flex;
   justify-content: center;
   align-items: center;
-  height: 30rem;
+  height: 20rem;
   width: 100%;
 }
