+++ src/styles/_variables.scss
@@ -27,7 +27,7 @@ $spacing-xl: 2rem;
 $border-radius-xsm: 0.1rem;
 $border-radius-sm: 0.25rem;
 $border-radius-md: 0.5rem;
-$border-radius-lg: 0.75rem;
+$border-radius-lg: 1.25rem;
 
 // Shadows
 $box-shadow-sm: 0 0.125rem 0.25rem rgba(0, 0, 0, 0.1);
